Game Manager, Cinema Controller, and Preload Scene (#6)
This commit adds a game manager, cinema controller, and sets up a Preloader Scene.

Game Manager is made as a singleton that will not be destroyed w/o specifically calling Destroy() on it.
Settings are parameterized so that we can modify on the fly either in the scene (locally) or in the prefab (globally).

Cinema Controller can be used to control cutscenes of any type before transitioning to the boss fights themselves.

The preload scene will be responsible for pre-loading any resources necessary as well as creating any "manager" singleton classes. (Non-singletons should just be created normally and not done this way - we also shouldn't make this a habit, just use non-singletons for most cases.)